+++ source/DSCResources/MSFT_SqlServerNetwork/MSFT_SqlServerNetwork.psm1
@@ -4,7 +4,7 @@
     THIS RESOURCE IS DEPRECATED!
 
     Changes to this resource will no longer be merged. Instead please use the
-    resources SqlServerProtocol and SqlServerProtocolTcpIp.
+    resources SqlProtocol and SqlProtocolTcpIp.
 #>
 $script:sqlServerDscHelperModulePath = Join-Path -Path $PSScriptRoot -ChildPath '..\..\Modules\SqlServerDsc.Common'
 $script:resourceHelperModulePath = Join-Path -Path $PSScriptRoot -ChildPath '..\..\Modules\DscResource.Common'