+++ source/Examples/Resources/SqlProtocol/1-EnableTcpIp.ps1
@@ -1,6 +1,7 @@
 <#
     .DESCRIPTION
-        This example will disable the Shared Memory protocol.
+        This example will enable the TCP/IP protocol, set the protocol to listen
+        on all IP addresses, and set the keep alive duration.
 
         The resource will be run as the account provided in $SystemAdministratorAccount.
 #>
@@ -17,11 +18,13 @@ Configuration Example
 
     node localhost
     {
-        SqlServerProtocol 'ChangeTcpIpOnDefaultInstance'
+        SqlProtocol 'ChangeTcpIpOnDefaultInstance'
         {
             InstanceName           = 'MSSQLSERVER'
-            ProtocolName           = 'SharedMemory'
-            Enabled                = $false
+            ProtocolName           = 'TcpIp'
+            Enabled                = $true
+            ListenOnAllIpAddresses = $false
+            KeepAlive              = 20000
 
             PsDscRunAsCredential   = $SystemAdministratorAccount
         }