+++ source/Examples/Resources/SqlProtocol/4-DisableTcpIp.ps1
@@ -1,7 +1,6 @@
 <#
     .DESCRIPTION
-        This example will enable the TCP/IP protocol, set the protocol to listen
-        on all IP addresses, and set the keep alive duration.
+        This example will disable the TCP/IP protocol.
 
         The resource will be run as the account provided in $SystemAdministratorAccount.
 #>
@@ -18,13 +17,11 @@ Configuration Example
 
     node localhost
     {
-        SqlServerProtocol 'ChangeTcpIpOnDefaultInstance'
+        SqlProtocol 'ChangeTcpIpOnDefaultInstance'
         {
             InstanceName           = 'MSSQLSERVER'
             ProtocolName           = 'TcpIp'
-            Enabled                = $true
-            ListenOnAllIpAddresses = $false
-            KeepAlive              = 20000
+            Enabled                = $false
 
             PsDscRunAsCredential   = $SystemAdministratorAccount
         }