+++ tests/Integration/DSC_SqlProtocolTcpIp.Integration.Tests.ps1
@@ -6,7 +6,7 @@ if (-not (Test-BuildCategory -Type 'Integration' -Category @('Integration_SQL201
 }
 
 $script:dscModuleName = 'SqlServerDsc'
-$script:dscResourceFriendlyName = 'SqlServerProtocolTcpIp'
+$script:dscResourceFriendlyName = 'SqlProtocolTcpIp'
 $script:dscResourceName = "DSC_$($script:dscResourceFriendlyName)"
 
 try